just some refactoring. figured it'd be more pythonic to use a mother object as qsiprep manager
